+++ com.unity.cinemachine/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - ThirdPerson Shooter sample scene now has an option to swap shoulders.
 
 ### Changed
+- Improve handling of nested blends
 
 
 ## [3.0.0-pre.5] - 2023-04-25